CDEC-461: Invert dependency between block and db packages
