+++ block/src/Pos/Block/Logic.hs
@@ -1,16 +1,6 @@
 -- | This module re-exports everything from 'Pos.Block.Logic.*'.
 module Pos.Block.Logic
-       ( module Pos.Block.Logic.VAR
-       , module Pos.Block.Logic.Util
-       , module Pos.Block.Logic.Internal
-       , module Pos.Block.Logic.Header
-       , module Pos.Block.Logic.Creation
-       , module Pos.Block.Logic.Integrity
+       ( module Pos.Block.Logic.Integrity
        ) where
 
-import           Pos.Block.Logic.Creation
-import           Pos.Block.Logic.Header
 import           Pos.Block.Logic.Integrity
-import           Pos.Block.Logic.Internal
-import           Pos.Block.Logic.Util
-import           Pos.Block.Logic.VAR
