+++ block/src/Pos/Block/Slog.hs
@@ -3,11 +3,7 @@
 -- another part of 'Pos.Block' is dedicated to whole blocks.
 
 module Pos.Block.Slog
-       ( module Pos.Block.Slog.Context
-       , module Pos.Block.Slog.Logic
-       , module Pos.Block.Slog.Types
+       ( module Pos.Block.Slog.Types
        ) where
 
-import           Pos.Block.Slog.Context
-import           Pos.Block.Slog.Logic
 import           Pos.Block.Slog.Types
