+++ explorer/src/Pos/Explorer/BListener.hs
@@ -29,7 +29,6 @@ import qualified Ether
 import           System.Wlog (WithLogger)
 import           UnliftIO (MonadUnliftIO)
 
-import           Pos.Block.BListener (MonadBListener (..))
 import           Pos.Block.Types (Blund)
 import           Pos.Core (HasConfiguration, HeaderHash, LocalSlotIndex (..),
                      SlotId (..), difficultyL, epochIndexL, getChainDifficulty,
@@ -40,6 +39,7 @@ import           Pos.Core.Chrono (NE, NewestFirst (..), OldestFirst (..),
 import           Pos.Core.Txp (Tx, txpTxs)
 import           Pos.Crypto (withHash)
 import           Pos.DB.BatchOp (SomeBatchOp (..))
+import           Pos.DB.Block (MonadBListener (..))
 import           Pos.DB.Class (MonadDBRead)
 import           Pos.Explorer.DB (Epoch, EpochPagedBlocksKey, Page,
                      defaultPageSize, findEpochMaxPages, numOfLastTxs)
