+++ generator/src/Test/Pos/Block/Logic/Event.hs
@@ -18,11 +18,11 @@ import           Control.Exception.Safe (fromException)
 import qualified Data.Map as Map
 import qualified Data.Text as T
 
-import           Pos.Block.Logic.VAR (BlockLrcMode, rollbackBlocks,
-                     verifyAndApplyBlocks)
 import           Pos.Block.Types (Blund)
 import           Pos.Core (HasConfiguration, HeaderHash)
 import           Pos.Core.Chrono (NE, NewestFirst, OldestFirst)
+import           Pos.DB.Block (BlockLrcMode, rollbackBlocks,
+                     verifyAndApplyBlocks)
 import           Pos.DB.Pure (DBPureDiff, MonadPureDB, dbPureDiff, dbPureDump,
                      dbPureReset)
 import           Pos.DB.Txp (MonadTxpLocal)
