+++ generator/test/Test/Pos/Block/Logic/CreationSpec.hs
@@ -17,7 +17,6 @@ import           Test.QuickCheck (Gen, Property, Testable, arbitrary, choose,
 
 
 import           Pos.Binary.Class (biSize)
-import           Pos.Block.Logic (RawPayload (..), createMainBlockPure)
 import qualified Pos.Communication ()
 import           Pos.Core (BlockVersionData (bvdMaxBlockSize), HasConfiguration,
                      SlotId (..), blkSecurityParam, genesisBlockVersionData,
@@ -28,6 +27,7 @@ import           Pos.Core.Ssc (SscPayload (..))
 import           Pos.Core.Txp (TxAux)
 import           Pos.Core.Update (UpdatePayload (..))
 import           Pos.Crypto (SecretKey)
+import           Pos.DB.Block (RawPayload (..), createMainBlockPure)
 import           Pos.Delegation (DlgPayload, ProxySKBlockInfo)
 import           Pos.Ssc.Base (defaultSscPayload)
 import           Pos.Update.Configuration (HasUpdateConfiguration)
