+++ lib/cardano-sl.cabal
@@ -310,6 +310,7 @@ test-suite cardano-test
                      , cardano-sl
                      , cardano-sl-binary
                      , cardano-sl-binary-test
+                     , cardano-sl-block
                      , cardano-sl-block-test
                      , cardano-sl-core
                      , cardano-sl-core-test
