+++ lib/src/Pos/GState.hs
@@ -6,16 +6,16 @@ module Pos.GState
        ( module Pos.DB.GState.Stakes
        , module Pos.DB.GState.Common
        , module Pos.DB.Delegation
-       , module Pos.GState.BlockExtra
+       , module Pos.DB.Block
        , module Pos.GState.Context
        , module Pos.GState.GState
        , module Pos.DB.Update
        ) where
 
+import           Pos.DB.Block
 import           Pos.DB.Delegation
 import           Pos.DB.GState.Common
 import           Pos.DB.GState.Stakes
 import           Pos.DB.Update
-import           Pos.GState.BlockExtra
 import           Pos.GState.Context
 import           Pos.GState.GState
