+++ lib/src/Pos/GState/GState.hs
@@ -7,6 +7,7 @@ module Pos.GState.GState
 import           Universum
 
 import           Pos.Core (GenesisData (..), HeaderHash, genesisData)
+import           Pos.DB.Block (initGStateBlockExtra)
 import           Pos.DB.Class (MonadDB)
 import           Pos.DB.Delegation (initGStateDlg)
 import           Pos.DB.GState.Common (initGStateCommon, isInitialized,
@@ -14,7 +15,6 @@ import           Pos.DB.GState.Common (initGStateCommon, isInitialized,
 import           Pos.DB.Ssc (initSscDB)
 import           Pos.DB.Txp (initGStateStakes, initGStateUtxo)
 import           Pos.DB.Update (initGStateUS)
-import           Pos.GState.BlockExtra (initGStateBlockExtra)
 import           Pos.Txp.GenesisUtxo (genesisUtxo)
 
 -- | Put missing initial data into GState DB.
