+++ lib/src/Pos/Logic/Types.hs
@@ -17,8 +17,6 @@ import           Data.Tagged (Tagged)
 import           Pipes (Producer)
 import           Pipes.Internal (unsafeHoist)
 
-import           Pos.Block.Logic (GetHashesRangeError,
-                     GetHeadersFromManyToError)
 import           Pos.Communication (NodeId)
 import           Pos.Core (HeaderHash, ProxySKHeavy, StakeholderId)
 import           Pos.Core.Block (Block, BlockHeader)
@@ -26,6 +24,7 @@ import           Pos.Core.Chrono (NE, NewestFirst (..), OldestFirst (..))
 import           Pos.Core.Txp (TxId, TxMsgContents)
 import           Pos.Core.Update (BlockVersionData, UpId, UpdateProposal,
                      UpdateVote, VoteId)
+import           Pos.DB.Block (GetHashesRangeError, GetHeadersFromManyToError)
 import           Pos.DB.Class (SerializedBlock)
 import           Pos.Security.Params (SecurityParams (..))
 import           Pos.Ssc.Message (MCCommitment, MCOpening, MCShares,
