+++ lib/src/Pos/Worker/Block.hs
@@ -23,14 +23,10 @@ import           System.Wlog (logDebug, logError, logInfo, logWarning)
 import           Pos.Block.Configuration (HasBlockConfiguration, criticalCQ,
                      criticalCQBootstrap, fixedTimeCQSec, networkDiameter,
                      nonCriticalCQ, nonCriticalCQBootstrap)
-import           Pos.Block.Logic (calcChainQualityFixedTime, calcChainQualityM,
-                     calcOverallChainQuality, createGenesisBlockAndApply,
-                     createMainBlockAndApply)
 import           Pos.Block.Slog (scCQFixedMonitorState, scCQOverallMonitorState,
                      scCQkMonitorState, scCrucialValuesLabel,
                      scDifficultyMonitorState, scEpochMonitorState,
-                     scGlobalSlotMonitorState, scLocalSlotMonitorState,
-                     slogGetLastSlots)
+                     scGlobalSlotMonitorState, scLocalSlotMonitorState)
 import           Pos.Core (BlockVersionData (..), ChainDifficulty, FlatSlotId,
                      HasProtocolConstants, SlotId (..), Timestamp (Timestamp),
                      addressHash, blkSecurityParam, difficultyL,
@@ -44,6 +40,9 @@ import           Pos.Core.Reporting (HasMisbehaviorMetrics, MetricMonitor (..),
                      MetricMonitorState, noReportMonitor, recordValue)
 import           Pos.Crypto (ProtocolMagic, ProxySecretKey (pskDelegatePk))
 import           Pos.DB (gsIsBootstrapEra)
+import           Pos.DB.Block (calcChainQualityFixedTime, calcChainQualityM,
+                     calcOverallChainQuality, createGenesisBlockAndApply,
+                     createMainBlockAndApply, slogGetLastSlots)
 import qualified Pos.DB.BlockIndex as DB
 import           Pos.DB.Delegation (getDlgTransPsk, getPskByIssuer)
 import qualified Pos.DB.Lrc as LrcDB (getLeadersForEpoch)
