+++ lib/test/Test/Pos/Block/CborSpec.hs
@@ -9,6 +9,7 @@ import           Universum
 import           Test.Hspec (Spec, describe)
 import           Test.Hspec.QuickCheck (modifyMaxSuccess)
 
+import qualified Pos.Block.Types as Block
 import           Pos.Core.Configuration (defaultCoreConfiguration,
                      withGenesisSpec)
 import qualified Pos.Network.Block.Types as Block
@@ -16,11 +17,11 @@ import qualified Pos.Network.Block.Types as Block
 import           Test.Pos.Binary.Helpers (binaryTest)
 import           Test.Pos.Block.Arbitrary.Message ()
 import           Test.Pos.Core.Arbitrary ()
-
+import           Test.Pos.DB.Block.Arbitrary ()
 
 spec :: Spec
 spec =
-    withGenesisSpec 0 defaultCoreConfiguration $ \_ ->
+    withGenesisSpec 0 defaultCoreConfiguration $ \_ -> do
         describe "Block network types" $ modifyMaxSuccess (min 10) $ do
             binaryTest @Block.MsgGetHeaders
             binaryTest @Block.MsgGetBlocks
@@ -28,3 +29,10 @@ spec =
             binaryTest @Block.MsgBlock
             binaryTest @Block.MsgStream
             binaryTest @Block.MsgStreamBlock
+
+        describe "Block types defined in the block package" $ do
+            describe "Bi instances" $ do
+                describe "Undo" $ do
+                    binaryTest @Block.SlogUndo
+                    modifyMaxSuccess (min 50) $ do
+                        binaryTest @Block.Undo
