+++ tools/src/blockchain-analyser/Main.hs
@@ -14,7 +14,7 @@ import           Pos.Core.Block (Block)
 import           Pos.Core.Chrono (NewestFirst (..))
 import           Pos.DB (closeNodeDBs, openNodeDBs)
 import           Pos.DB.Block (getUndo)
-import qualified Pos.DB.Block.Load as DB
+import qualified Pos.DB.Block as DB
 import           Pos.DB.Class (getBlock)
 import qualified Pos.DB.GState.Common as GS
 import           Pos.Launcher (withConfigurations)
